requirements.txt
numpy==1.26.4
openai==1.34.0
opencv-python==4.10.0.82
pillow==10.3.0
requests==2.32.3
RPi.GPIO==0.7.1
